+++ packages/lambda-postgres-query/src/cloudformation/index.ts
@@ -1,3 +1,2 @@
 export * from './createLambdaPostgresQueryTemplate';
 export * from './lambdaQueryHandler';
-export * from './lambdaReadOnlyQueryHandler';