+++ packages/ui/tsup.config.ts
@@ -3,4 +3,9 @@ import { tsupConfig } from '@ttoss/config';
 export const tsup = {
   ...tsupConfig(),
   format: ['esm'],
+  dts: {
+    compilerOptions: {
+      noCheck: true,
+    },
+  },
 };