+++ packages/appsync-api/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.24.5](https://github.com/ttoss/ttoss/compare/@ttoss/appsync-api@0.24.4...@ttoss/appsync-api@0.24.5) (2026-05-14)
+
+**Note:** Version bump only for package @ttoss/appsync-api
+
+## [0.24.4](https://github.com/ttoss/ttoss/compare/@ttoss/appsync-api@0.24.3...@ttoss/appsync-api@0.24.4) (2026-05-13)
+
+**Note:** Version bump only for package @ttoss/appsync-api
+
+## [0.24.3](https://github.com/ttoss/ttoss/compare/@ttoss/appsync-api@0.24.2...@ttoss/appsync-api@0.24.3) (2026-05-13)
+
+**Note:** Version bump only for package @ttoss/appsync-api
+
 ## [0.24.2](https://github.com/ttoss/ttoss/compare/@ttoss/appsync-api@0.24.1...@ttoss/appsync-api@0.24.2) (2026-05-11)
 
 ### Bug Fixes