+++ packages/cloud-vpc/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.45](https://github.com/ttoss/ttoss/compare/@ttoss/cloud-vpc@0.1.44...@ttoss/cloud-vpc@0.1.45) (2026-05-14)
+
+**Note:** Version bump only for package @ttoss/cloud-vpc
+
+## [0.1.44](https://github.com/ttoss/ttoss/compare/@ttoss/cloud-vpc@0.1.43...@ttoss/cloud-vpc@0.1.44) (2026-05-13)
+
+**Note:** Version bump only for package @ttoss/cloud-vpc
+
+## 0.1.43 (2026-05-13)
+
+### Bug Fixes
+
+- **forms:** map field.value to checked prop in FormFieldSwitch ([#981](https://github.com/ttoss/ttoss/issues/981)) ([9581d66](https://github.com/ttoss/ttoss/commit/9581d66470293699f775009698e5cc1d0eaeeee1))
+
 ## [0.1.42](https://github.com/ttoss/ttoss/compare/@ttoss/cloud-vpc@0.1.41...@ttoss/cloud-vpc@0.1.42) (2026-04-23)
 
 **Note:** Version bump only for package @ttoss/cloud-vpc