+++ packages/fsl-theme/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.12](https://github.com/ttoss/ttoss/compare/@ttoss/fsl-theme@1.1.11...@ttoss/fsl-theme@1.1.12) (2026-05-13)
+
+**Note:** Version bump only for package @ttoss/fsl-theme
+
 ## 1.1.11 (2026-05-12)
 
 **Note:** Version bump only for package @ttoss/fsl-theme