+++ packages/graphql-api-server/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.11.14 (2026-05-13)
+
+### Bug Fixes
+
+- **forms:** map field.value to checked prop in FormFieldSwitch ([#981](https://github.com/ttoss/ttoss/issues/981)) ([9581d66](https://github.com/ttoss/ttoss/commit/9581d66470293699f775009698e5cc1d0eaeeee1))
+
 ## [0.11.13](https://github.com/ttoss/ttoss/compare/@ttoss/graphql-api-server@0.11.12...@ttoss/graphql-api-server@0.11.13) (2026-04-28)
 
 **Note:** Version bump only for package @ttoss/graphql-api-server