+++ packages/lambda-postgres-query/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/ttoss/ttoss/compare/@ttoss/lambda-postgres-query@1.1.1...@ttoss/lambda-postgres-query@1.1.2) (2026-05-14)
+
+### Bug Fixes
+
+- lambda function name ([cd2fd63](https://github.com/ttoss/ttoss/commit/cd2fd631c40572d2953e4baada349dc24170c607))
+
+## [1.1.1](https://github.com/ttoss/ttoss/compare/@ttoss/lambda-postgres-query@1.1.0...@ttoss/lambda-postgres-query@1.1.1) (2026-05-14)
+
+**Note:** Version bump only for package @ttoss/lambda-postgres-query
+
+# [1.1.0](https://github.com/ttoss/ttoss/compare/@ttoss/lambda-postgres-query@1.0.1...@ttoss/lambda-postgres-query@1.1.0) (2026-05-13)
+
+### Features
+
+- lambda postgres query security ([#996](https://github.com/ttoss/ttoss/issues/996)) ([dff9948](https://github.com/ttoss/ttoss/commit/dff9948e3b9f6d83192b1c26df227dfc7341638d))
+
+## 1.0.1 (2026-05-13)
+
+### Bug Fixes
+
+- **forms:** map field.value to checked prop in FormFieldSwitch ([#981](https://github.com/ttoss/ttoss/issues/981)) ([9581d66](https://github.com/ttoss/ttoss/commit/9581d66470293699f775009698e5cc1d0eaeeee1))
+
 ## [0.6.2](https://github.com/ttoss/ttoss/compare/@ttoss/lambda-postgres-query@0.6.1...@ttoss/lambda-postgres-query@0.6.2) (2026-04-23)
 
 **Note:** Version bump only for package @ttoss/lambda-postgres-query