+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.9.18](https://github.com/ttoss/ttoss/compare/@ttoss/ui@6.9.17...@ttoss/ui@6.9.18) (2026-05-13)
+
+**Note:** Version bump only for package @ttoss/ui
+
 ## [6.9.17](https://github.com/ttoss/ttoss/compare/@ttoss/ui@6.9.16...@ttoss/ui@6.9.17) (2026-05-13)
 
 **Note:** Version bump only for package @ttoss/ui